split out helper functions to seperate file and added missings tests for database manager
